requirements.txt
alembic==1.6.3
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
dnspython==2.1.0
email-validator==1.1.2
Flask==2.0.1
Flask-Dance==5.0.0
Flask-Login==0.5.0
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
oauthlib==3.1.0
pycodestyle==2.7.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.15
toml==0.10.2
urllib3==1.26.4
URLObject==2.4.3
Werkzeug==2.0.1
WTForms==2.3.3
